Added Tele Percent out
Goal is for the Spark max's to use percent output to lift the robot up until quite clost to the bar. Then switch to PID control, which will align the arms with the setpoint using less power and a .1 or .2 KF in stead of a .4 or a .5 KF. The idea is to reduce power usage by the neo's, to reduce heat and power usage to prevent browning out, which may cause the motors to release and drop the robot.