Changed Climber sequence
Switched climber sequence to command from sequenced command. Is a bit more in the command file but should work like competition and move through steps when button is held down, when rotator's / tele's are in tollerence. FYI: Not all steps have tolerence checks, for both tele and arms. Ex. Base setup to lifting arms to 2nd bar, if rotator's are a degree or two off tollerence, it does not matter. But when the robot lifts onto 2 the tollerence is checked when the rotators lock, beacuse it is a critical point they need to be in tolerence.